// quasar-project/src/types/IMember.ts
export interface IMember {
  post: string | null;
  bid: string | null;
  salary: string | null;
  PKG: string | null;
  KL: string | null;
  postType: string | null;
  nameOfOrganization: string | null;
  subdivision: string | null;
  supervisor: string | null;
  signOfALeader: string | null;
  contractEndDate: string | null;
  status: string | null;
  reasonForStatusChange: string | null;
}